Extract Simple Experiment Details Component
