+++ pages/experiments/[id].tsx
@@ -4,56 +4,12 @@ import { toIntOrNull } from 'qc-to_int'
 import React, { useEffect, useState } from 'react'
 
 import ExperimentsApi from '@/api/ExperimentsApi'
+import ExperimentDetails from '@/components/ExperimentDetails'
 import ExperimentTabs from '@/components/ExperimentTabs'
 import Layout from '@/components/Layout'
 import { ExperimentFull } from '@/models'
-import { formatIsoUtcOffset } from '@/utils/formatters'
 
 const debug = debugFactory('abacus:pages/experiments/[id].tsx')
-
-function ExperimentDetails({ experiment }: { experiment: ExperimentFull }) {
-  return (
-    <div>
-      <h2>Experiment details</h2>
-      <table>
-        <tbody>
-          <tr>
-            <td>Name</td>
-            <td>{experiment.name}</td>
-          </tr>
-          <tr>
-            <td>P2 Link</td>
-            <td>
-              <a href={experiment.p2Url} rel='noopener noreferrer' target='_blank'>
-                P2
-              </a>
-            </td>
-          </tr>
-          <tr>
-            <td>Description</td>
-            <td>{experiment.description}</td>
-          </tr>
-          <tr>
-            <td>Start</td>
-            <td>{formatIsoUtcOffset(experiment.startDatetime)}</td>
-          </tr>
-          <tr>
-            <td>End</td>
-            <td>{formatIsoUtcOffset(experiment.endDatetime)}</td>
-          </tr>
-          <tr>
-            <td>Status</td>
-            <td>{experiment.status}</td>
-          </tr>
-          <tr>
-            <td>Platform</td>
-            <td>{experiment.platform}</td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  )
-}
 
 export default function ExperimentPage() {
   const experimentId = toIntOrNull(useRouter().query.id)
